Unify cvar commands
